configurator ui

+++ configurator/src/main/java/io/bytebeam/uplink/service/UplinkService.java
@@ -2,6 +2,7 @@ package io.bytebeam.uplink.service;
 
 import android.app.Service;
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.*;
 import android.util.Log;
 import androidx.annotation.Nullable;
@@ -16,39 +17,80 @@ import java.util.List;
 import static io.bytebeam.uplink.common.Constants.*;
 
 public class UplinkService extends Service {
-    public static String TAG = "UplinkService";
+    public static final String TAG = "UplinkService";
+    public static final String KILL_COMMAND_TAG = "KILL_SERVICE";
+    public static final String PREFS_NAME = "configuration";
+    public static final String PREFS_AUTH_CONFIG_KEY = "auth_config";
+    public static final String PREFS_AUTH_CONFIG_NAME_KEY = "auth_config_name";
+    public static final String PREFS_CLIENTS_COUNT_KEY = "clients_count";
     List<Messenger> subscribers = new ArrayList<>();
     long uplink = 0;
+
+    int clientsCount = 0;
+    private void setClientsCount(int newCount) {
+        clientsCount = newCount;
+        SharedPreferences.Editor prefs = getApplicationContext().getSharedPreferences(PREFS_NAME, MODE_PRIVATE).edit();
+        prefs.putInt(PREFS_CLIENTS_COUNT_KEY, clientsCount);
+        prefs.apply();
+    }
 
     @Nullable
     @Override
     public IBinder onBind(Intent intent) {
-        uplink = NativeApi.createUplink(
-                Utils.getRawTextFile(getApplicationContext(), R.raw.auth_config),
-                String.format(
-                        "[persistence]\n" +
-                                "path = \"%s/uplink\"\n" +
-                                "\n" +
-                                "[streams.battery_stream]\n" +
-                                "topic = \"/tenants/{tenant_id}/devices/{device_id}/events/battery_level\"\n" +
-                                "buf_size = 1\n",
-                        getApplicationContext().getFilesDir().getAbsolutePath()
-                ),
-                true,
-                this::uplinkSubscription
-        );
-        Log.d(TAG, "Uplink service created");
-        return new Messenger(new Handler(Looper.myLooper(), this::handleMessage)).getBinder();
+        if (intent.getBooleanExtra(KILL_COMMAND_TAG, false)) {
+            Log.d(TAG, "killing service");
+            onUnbind(null);
+            return null;
+        }
+
+        SharedPreferences prefs = getApplicationContext().getSharedPreferences(PREFS_NAME, MODE_PRIVATE);
+        String authConfig = prefs.getString(PREFS_AUTH_CONFIG_KEY, null);
+        if (authConfig == null) {
+            // TODO: handle this
+            authConfig = Utils.getRawTextFile(getApplicationContext(), R.raw.auth_config);
+        }
+
+        if (uplink == 0) {
+            uplink = NativeApi.createUplink(
+                    authConfig,
+                    String.format(
+                            "[persistence]\n" +
+                                    "path = \"%s/uplink\"\n" +
+                                    "\n" +
+                                    "[streams.battery_stream]\n" +
+                                    "topic = \"/tenants/{tenant_id}/devices/{device_id}/events/battery_level\"\n" +
+                                    "buf_size = 1\n",
+                            getApplicationContext().getFilesDir().getAbsolutePath()
+                    ),
+                    true,
+                    this::uplinkSubscription
+            );
+            Log.d(TAG, "uplink native context initialized");
+        }
+
+        IBinder result = new Messenger(new Handler(Looper.myLooper(), this::handleMessage)).getBinder();
+
+        try {
+            result.linkToDeath(() -> {
+                Log.d(TAG, "binder died");
+                setClientsCount(clientsCount-1);
+            }, 0);
+        } catch (RemoteException e) {}
+        setClientsCount(clientsCount+1);
+
+        return result;
     }
 
     @Override
     public boolean onUnbind(Intent intent) {
         Log.d(TAG, "shutting down uplink service and process");
         subscribers.clear();
-        NativeApi.destroyUplink(uplink);
-        uplink = 0;
+        if (uplink != 0) {
+            NativeApi.destroyUplink(uplink);
+            uplink = 0;
+        }
         // forcefully kill the service process to allow the cleanup of the native resources
-        Runtime.getRuntime().exit(0);
+        System.exit(0);
         return false;
     }
